project02-main/requirements.txt
flask
flask-socketio
flask-wtf
python-dotenv
pymysql
